refactor : 티켓팅 페이지 컴포넌트 분리
